src/pyproject.toml
[project]
name = "credit-api"
version = "0.1.0"
description = "Credit Risk API (Django)"
requires-python = ">=3.11"

dependencies = [
  "django==4.1.5",
  "psycopg2-binary",
  "djangorestframework",
  "drf-spectacular",
  "django-safedelete",
  "factory_boy",
]

[tool.uv]
dev-dependencies = [
  "black",
  "ruff",
  "pytest",
  "pytest-django",
]